% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_class.R
\docType{class}
\name{advice}
\alias{advice}
\alias{advice-class}
\alias{class:advice}
\alias{advice-method}
\alias{movie_title<-}
\alias{show,advice-method}
\alias{movie_title,advice-method}
\alias{movie_title<-,advice-method}
\title{An S4 class to represent a movie advice}
\usage{
\S4method{show}{advice}(object)

\S4method{movie_title}{advice}(object)

\S4method{movie_title}{advice}(object) <- value

advice(movie_title, plot = NA_character_, keywords = NA_character_)
}
\arguments{
\item{object}{an object of class advice}

\item{value}{value to assign to the movie title}

\item{movie_title}{movie title}

\item{plot}{plot}

\item{keywords}{keywords}
}
\value{
An \code{\link{advice-class}}
}
\description{
An S4 class to represent a movie advice
}
\section{Functions}{
\itemize{
\item \code{show(advice)}: show the advice

\item \code{movie_title(advice)}: movie_title of the advice

\item \code{movie_title(advice) <- value}: movie_title of the advice

}}
\examples{
new_adv <- advice("title1")
new_adv
# advice(32) # not working (internal validation)
}
\seealso{
\code{\link{advice}}
}
