% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise.R
\name{advise.good.movie}
\alias{advise.good.movie}
\title{Advise movies based on another movie}
\usage{
advise.good.movie(similar_to, how_many, ...)
}
\arguments{
\item{similar_to}{character, movie title from the database.}

\item{how_many}{integer, how many movies to advise.}

\item{...}{additional arguments.}
}
\description{
Documentation...
}
