% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise_movie.R
\name{advise_movie}
\alias{advise_movie}
\title{Advise movies based on another movie}
\usage{
advise_movie(similar_to, how_many, draw_scores = FALSE, ...)
}
\arguments{
\item{similar_to}{character, movie title from the database.}

\item{how_many}{integer, how many movies to advise.}

\item{draw_scores}{if`TRUE`, draws a barplot with the similarity scores.}

\item{...}{additional arguments (for now, only`weights`).}
}
\value{
A`list` with the following elements:`movie_title`: advised movie title(s),
`plot`: plot(s) of advised movie title(s),
`movie_ids`: line number in the`movies_DB` database of the advised movie(s),
`scores`: similarity scores of the advised movie(s)
}
\description{
`advise_movie` takes as input a movie from the movie database and gives as output a number
of movies that are similar.
}
\examples{
suggestions <- advise_movie(similar_to = "Interstellar", how_many = 3,
draw_scores = TRUE, weights = c("genre"=1, "popularity"=1, "rating"=1,
"production company"=1))

}
