% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise_movie_k.R
\name{advise_movie_k}
\alias{advise_movie_k}
\title{title}
\usage{
advise_movie_k(keywords, how_many)
}
\arguments{
\item{keywords}{bla}

\item{how_many}{bla}
}
\value{
list
}
\description{
description
}
\examples{
advise_movie_k(keywords = "christmas", how_many = 2)

}
