% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviesdemo-package.R, R/moviesdemo.R
\docType{package}
\name{moviesdemo-package}
\alias{moviesdemo-package}
\title{moviesdemo: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.

A demo package for the SMCS course on creating and publishing R packages,
based on metadata of 4800 movies from The Movie Database.
}
\section{moviesdemo functions}{

\describe{
  \item{\code{advise_movie}}{Advise movies based on another movie}
  \item{\code{advise_movie_k}}{Advise movies based on keywords}
}


\describe{
  \item{\code{advise_movie}}{Advise movies based on another movie}
  \item{\code{advise_movie_k}}{Advise movies based on keywords}
}
}

\section{Database}{

The database used is the TMDb movies Database (see \code{\link{movies_DB}})


The database used is the TMDb movies Database (see \code{\link{movies_DB}})
}

\keyword{4800}
\keyword{A}
\keyword{Database.}
\keyword{Movie}
\keyword{R}
\keyword{SMCS}
\keyword{The}
\keyword{and}
\keyword{based}
\keyword{course}
\keyword{creating}
\keyword{demo}
\keyword{for}
\keyword{from}
\keyword{internal}
\keyword{metadata}
\keyword{movies}
\keyword{of}
\keyword{on}
\keyword{package}
\keyword{packages,}
\keyword{publishing}
\keyword{the}
