% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviesdemo.R
\docType{package}
\name{moviesdemo}
\alias{moviesdemo}
\alias{moviesdemo-package}
\title{moviesdemo: A demo R package.}
\description{
A demo package for the SMCS course on creating and publishing R packages,
based on metadata of 4800 movies from The Movie Database.
}
\section{moviesdemo functions}{

\describe{
  \item{\code{advise.good.movie}}{Advise movies based on another movie}
  \item{\code{sim.genres}}{Movie similarity based on production companies}
  \item{\code{sim.producers}}{Movie similarity based on production companies}
}
}

\section{Shiny App}{

A Shiny App is available through the \code{\link{runMovieApp}} function
}

\section{Database}{

The database used is the TMDb 4800 movies Database (see \code{\link{movies}})
}

