% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{openSlides}
\alias{openSlides}
\title{Open html slides SMCS course}
\usage{
openSlides()
}
\description{
This function opens the slides of the SMCS course on creating and publishing
your R package with RStudio.
}
