% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sim.producers}
\alias{sim.producers}
\title{Movie similarity based on production companies}
\usage{
sim.producers(movie_A, movie_B)
}
\arguments{
\item{movie_A}{character, one movie title from the movie database}

\item{movie_B}{character, another movie title from the movie database}
}
\description{
Documentation...
}
