% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise.R
\name{advise.good.movie}
\alias{advise.good.movie}
\title{Advise movies based on another movie}
\usage{
advise.good.movie(similar_to, how_many, draw_scores = FALSE, ...)
}
\arguments{
\item{similar_to}{character, movie title from the database.}

\item{how_many}{integer, how many movies to advise.}

\item{draw_scores}{if \code{TRUE}, draws a barplot with the similarity scores.}

\item{...}{additional arguments.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{selected}}{movie title used to advise other movies}
  \item{\code{to_watch}}{advised movie title(s)}
  \item{\code{movie_ids}}{line number in the \code{movies} database of the advised movie(s)}
  \item{\code{scores}}{similarity scores of the advised movie(s)}
}
}
\description{
\code{advise.good.movie} takes as input a movie from the movie database and gives as output a number
of movies that are similar.
}
\examples{
suggestions <- advise.good.movie(similar_to = "Interstellar", how_many = 3,
draw_scores = TRUE, weights = c("genre"=1, "popularity"=1, "rating"=1,
"production company"=1))

}
