% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise_movie_k.R
\name{advise_movie_k}
\alias{advise_movie_k}
\title{Advise movies based on types}
\usage{
advise_movie_k(keywords, how_many)
}
\arguments{
\item{keywords}{list of keywords, Type of the movies}

\item{how_many}{integer, how many movies to advise.}
}
\value{
A`list` with the following elements:`movie_title`: advised movie title(s),
`plot`: plot(s) of advised movie title(s),
`keywords`: keywords of the new movies
}
\description{
`advise_movie_k` takes as input a list of movie genre gives as output a number
of movies that are in these genre
}
\examples{
suggestions <-advise_movie_k(c("princess", "magic"), how_many = 3)

}
