% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{TMDb 4800 movie dataset}
\format{An object of class \code{data.frame} with 4800 rows and 9 columns.}
\source{
\url{https://www.kaggle.com/tmdb/tmdb-movie-metadata}
}
\usage{
movies
}
\description{
Metadata on 4800 movies from TMDb.
}
\keyword{datasets}
