% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviesdemo.R
\docType{package}
\name{moviesdemo}
\alias{moviesdemo}
\alias{moviesdemo-package}
\title{moviesdemo: A demo R package.}
\description{
A demo package for the SMCS course on creating and publishing R packages,
based on metadata of 4800 movies from The Movie Database.
}
