% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sim.genres}
\alias{sim.genres}
\title{Movie similarity based on genre}
\usage{
sim.genres(movie_A, movie_B)
}
\arguments{
\item{movie_A}{character, one movie title from the movie database}

\item{movie_B}{character, another movie title from the movie database}
}
\value{
A scalar : the similarity based on genre
}
\description{
The similarity of 2 movies A and B is measured based on their
potential common genre.
}
